Rock 5B: Set combo PHY #0 to PCIE
On the Rock 5B, Combo PHYs #0 and #1 are used for PCIE, while #2 serves the top
USB 3.0 Type A port.

SATA is not accessible, except via a special M.2 breakout adapter, which connects to PHY #1.
Therefore, setting PHY #0 to SATA makes no sense. (In the future, PHY #1 should be made
configurable in the firmware menu, to enable using the M.2 -> SATA adapter.)