Rock 5B: Support SD card boot
Currently, this is done using a forked version of DwEmmcDxe, due to clock handling differences
between RK3568 and RK3588 (on RK3588, the SD interface clock is managed by SCMI).

The driver name is confusing - DwEmmcDxe is for the SD socket, while SdhciHostDxe is for the
eMMC socket. Unfortunately eMMC boot cannot be enabled, because SdhciHostDxe softlocks
if it cannot detect a plugged-in eMMC, instead of allowing fallback to another boot device.