Add config options for CPU clock rates and voltages
By default, clock rates are set to maximum and voltages are set depending on the clock.
This behavior can be changed in the CPU Performance form (for platforms that have persistent NV support as of now).

Clocks are set as soon as we get write access to the var store, however voltages will be set later on first notification of the ReadyToBoot event.
This allows the user to dial settings back in case the values are too high and the system hangs or boot loops when loading the OS.