Silicon/Rockchip: Support persistent variable store on SD/eMMC
On SD/eMMC this feature is more limited, since these devices are meant to be used by the OS and we can't take ownership of them.
As such, changes to the var store will only be saved at boot time when certain events happen (platform reset, ready to boot, image load).

Full runtime support is available when booted from SPI NOR flash.

Also introduce the gRockchipTokenSpaceGuid.PcdNvStoragePreferSpiFlash flag which forces the var store to be on SPI flash, if available, no matter the boot device.
This can be useful during development as NOR flash is quite slow to write compared to SD.