Platforms: Configure and enable PCIe support
It's been tested to work on all platforms, with the exception of Mekotronics boards:
PCIE20L0 on R58-Mini hangs at "PCIe: Enabling DBI access" and on R58X it doesn't detect the Wi-Fi chip.

This is likely happening due to incorrect pin muxing, since GMAC routing was causing similar issues on other boards.
Will be addressed in a future patch.