Silicon/RK3588: Fix PCIe addresses & add initial ACPI support
The configuration is almost identical to that of Quartz64-UEFI.
We hide the root port and expose only bus 1 to the OS due to non-compliant ECAM (therefore switches won't work).

ACPI support has been validated in Ubuntu Server 22.04.2 with NVME boot on PCIe 3x4 and onboard RTL8125B NIC on PCIe 2x1.

Windows is able to enumerate devices and even initialize the NVME drive, but unfortunately it keeps timing out.

Not yet enabled in platforms because we must expose only the controllers that are enabled/initialized.