Silicon/Rockchip: Introduce RockchipI2cMasterProtocol
We need a way to get the platform I2C bus number associated with the master protocol, so drivers that consume it directly (like RTC) can be provided with the appropriate controller handle.

RTC drivers don't follow the higher-level I2C I/O model as it doesn't support runtime use.