Initialize GMAC and expose in ACPI
Enabled on all platforms that use it.

Tested to be working with NetBSD on:
- R58X & R58-Mini
- Orange Pi 5
- ROCK 5A
- ROC-RK3588S-PC / Station M3

VMware ESXi also detects the NICs and link status, however it's not able to get
an IP address.

The original GMAC0 definition by Rockchip was more thorough, but I have no idea
if all the methods have been tested and with what kernel, so for now we'll keep
it minimal like on Quartz64.

This will only work when connected to a Gigabit capable port, lower speed modes
are not supported due to lack of clock management in the OS.