Add basic status LED support
During DXE phase, the LED will pulse repeatedly. If this keeps going on forever
or the LED freezes, something went terribly wrong.

After the console is ready, at BDS phase, the LED will switch to a calm short
pulsing every two seconds or so. This indicates the firmware is alive and
waiting for user input or the countdown to boot automatically. If the blinking
stops, it can either mean that the firmware has exited the boot services and
handed off control to the OS (most likely), or that it's not responding. Brief
freezes are expected with some peripherals due to how device drivers use the
timer.

I had initially planned on blinking the LED at ExitBootServices() too, but the
delay added to the boot process doesn't seem justified.

Long pulses >= 2 are reserved for future error code reporting.

Enabled on all platforms that have a controllable LED.