Revert "Fydetab Duo: fixes for SDcard"
With current change, SD card no longer appears in UEFI Boot Manager. Looking at commit history, the problem is already described and workaround:
https://github.com/edk2-porting/edk2-rk3588/commit/c02b333c22b4fff4569c6dcaa2b84696edae4ce3 (same as current change: bad)
https://github.com/edk2-porting/edk2-rk3588/commit/416b2fa11b031aa894bfccbc57e9be75388fdea1 (actual solution)

The Fydetab Duo motherboard has SDMMC_DET inverted, which is against Rockchip design. This is a hardware mistake that prevents the RK3588 MaskROM to boot from SD card when tray is fully closed. No software fix is possible because MaskROM is read-only. Therefore, in order to boot UEFI firmware and OS from SD card, the tray has to be left ajar and UEFI firmware must not invert the state once again.

This allows UEFI firmware to detect the SD card no matter if tray is fully in or partially.