Hook recovery key to MaskROM reset
If the recovery key is held right after power on, UEFI will reset to
MaskROM.

Some platforms don't have an easily accessible MaskROM button, but they
do have a recovery key (volume up on Fydetab) that can be used for this
purpose. U-Boot proper maps it to rockusb loader mode instead, however
we don't support this protocol.

This is done early in the SEC phase to reduce the chances of something
going horribly wrong (e.g. in DXE) and preventing recovery.

Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>